// MiLicenciaCentro/src/app/enums/API/MetodosCotizador.ts
/**
 * Métodos para el controlador de
 * Cotización
 */
 export enum MetodosCotizador {
    SeleccionarComerciosPorTipo = "SeleccionarComerciosPorTipo",
    NuevaVigenciaIp = "NuevaVigenciaIp",
    ConsultaIp = "ConsultaIp",
    Log = "Log",
    EnviarCorreo = "EnviarCorreos",
    EnviarCorreoCuota = "EnvioCorreoCuota",
    ObtenerTiposDeIdentificacion = "ObtenerTiposDeIdentificacion",
    ObtenerParametrosPinesOlimpia = 'ParametrosPinesOlimpia',
    ObtenerValorDelPin = 'Costo',
    ObtenerConveniosCentro = 'Convenio',
    ObtenerConveniosCentroCEA = 'ObtenerConveniosCentroCEA',
    ConsultaCentroActual = "ConsultaCentroActual",
    GenerarReferenciaPinesOlimpia = 'GenerarReferenciaBancolombia',
    ObtenerPlantillas = "ObtenerPlantillas",
    SeleccionarReferenciaTransaccionCuota = "SeleccionarReferenciaTransaccionCuota",
    ObtenerCategorias = 'ObtenerCategorias',
    ObtenerCentros = 'Centros',
    ConsultarPinesOlimpiaComprados = 'ConsultarPinesCompradosPSE',
    ConfirmarConsultaSecurePay = 'ConfirmarEstadoRecaudoConsultaSP',
    ConsultarPinesCEA = "ConsultaPinesCea",
    ObtenerListaBancos = 'ObtenerListaBancos',
    AccionesCambioBancolombia = "AccionesBeneficiarioBancolombia",
    ConsultaEstadoDevolucion = "ConsultaEstadoDevolucion",
    ConsultaEstadoCambioBeneficiario = "ConsultaEstadoCambioBeneficiario",
    ConsultaEstadoAgendamiento = "ConsultaEstadoAgendamiento",
    ConfirmarDevolucion = "ConfirmarDevolucion",
    GetAgenteDispersionByPin = "GetAgenteDispersionByPin",
    ConfirmarCambioDocumento = "ConfirmarCambioDocumento",
    DevolucionByDocumento = "devolucionByDocumento",
    seleccionarResumenPseByPin = "seleccionarResumenPseByPin",
    seleccionarResumenByPin = "seleccionarResumenByPin",
    PrevalidarActualizacionDevolucion = "prevalidarActualizacionDevolucion",
    ActualizacionDevolucion= "actualizacionDevolucion",
    ObtenerTipoDocumentos = "obtenerTipoDocumentos",
    ConsultarInformacionPinDevoluciones="consultarInformacionPinDevolucion",
    ActualizarContacto="actualizarContacto",
    GetMotivoDevolucionByTipoPin="getMotivoDevolucionByTipoPin",
    ActualizarDevolucionTransferencia="actualizarDevolucionTransferencia",
    GuardarCertificado="guardarCertificado",
    GetCertificadoByPin="getCertificadoByPin",
    ValidacionOtp = "validacionOtp",
    EnvioOtp = "envioOtp",
    PermiteFacturaElectronica ="CheckElectronicBillingStatus",
    ObtenerTipoPersona = "GetPersonType",
    ObtenerTipoDocumento = "GetIdDocumentType",
    ObtenerTokenAceptacion = "ObtenerTokenAceptacion",
    GenerarReferenciaBancolombiaWompi = "GenerarReferenciaBancolombiaWompiCaptcha",
    ConsultaExistePin = "ConsultaExistePin",
    ConstruirCorreoWompi = "ConstruirCorreoWompi",
    ObtenerCategoriasCentro="ConsultaCategoriasCentro",
    AutorizarPagoDaviplata = "AutorizarPagoDaviplata",
  }