// MiLicenciaCentro/src/app/enums/API/Rutas.ts
export enum Rutas{
    ServiciosCotizador = '/servicios/cotizador/',
    ServiciosCompraPin = '/servicios/CompraPin/',
    ServiciosPerfil = '/servicios/perfil/',
    ServiciosAgendamientoEnrolmiento = '/servicios/AgendamientoEnrolamiento/',
    ServicioSecurePay = '/servicios/SecurePay/',
    ServicioTuPago = '/servicios/TuPago/',
    ServicioCentro = '/servicios/Centro/',
}